chore: updated sqlalchemy scaffold structure
