+++ robyn/scaffold/structured/sqlalchemy/adaptors/models/__init__.py
@@ -1,0 +1,5 @@
+from user import Base as user
+
+metadata = [
+    user.metadata
+]
