+++ robyn/scaffold/structured/sqlalchemy/adaptors/models/user.py
@@ -1,6 +1,6 @@
 from sqlalchemy import String, Integer, Boolean
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
-
+from pydantic import BaseModel
 
 class Base(DeclarativeBase):
     pass
@@ -14,3 +14,22 @@ class User(Base):
     hashed_password: Mapped[str] = mapped_column(String)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     is_superuser: Mapped[bool] = mapped_column(Boolean, default=False)
+
+
+class UserRead(BaseModel):
+    id: int
+    username: str
+    is_active: bool
+    is_superuser: bool
+
+    class Config:
+        orm_mode = True
+
+
+# Pydantic model for creating/updating a user
+class UserCreate(BaseModel):
+    username: str
+    password: str
+
+    class Config:
+        orm_mode = True
